requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
diff-match-patch==20241021
Django==5.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-import-export==4.3.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
idna==3.10
pillow==11.3.0
PyJWT==2.10.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
urllib3==2.5.0
